refactor: query client provider를 분리
