+++ src/app/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
-import "./index.css"
+import "./style/index.css"
 import App from "@/app/App.tsx"
 
 createRoot(document.getElementById("root")!).render(
