+++ src/app/providers/QueryClientProvider.tsx
@@ -1,5 +1,5 @@
-import { QueryClient, QueryClientProvider as ReactQueryClientProvider } from "@tanstack/react-query"
 import { useState } from "react"
+import { QueryClient, QueryClientProvider as ReactQueryClientProvider } from "@tanstack/react-query"
 
 export const QueryClientProvider = ({ children }: { children: React.ReactNode }) => {
   const [queryClient] = useState(() => new QueryClient())
