+++ src/pages/PostsManagerPage.tsx
@@ -1,6 +1,6 @@
 import { Card, CardContent } from "@/shared/ui"
 
-import { UserModal } from "../components/UserModal"
+import { UserModal } from "@/components/dialog/UserModal"
 import { AddCommentDialog } from "../components/AddCommentDialog"
 import { EditPostDialog } from "../components/dialog/EditPostDialog"
 import { EditCommentDialog } from "../components/dialog/EditCommentDialog"
